Use a cache to store symbols rather than iterating over an array over an over.
Iterating over an array to check if an element is an operator is O(n) where n is the number of elements in the array.
If the method we're analyzing is really big the result of checking operators for each element can be really high.

Instead, we just count the amount of times a symbol shows up inside the method and sum the result for the operators we want.
Incrementing the cache and getting the results is O(1), which is much better.